+++ lib/responsive_design.dart
@@ -1,4 +1,5 @@
 export 'src/domain/entities/responsive_settings.dart';
+export 'src/domain/entities/screen_change_points.dart';
 export 'src/domain/enums/responsive_type.dart';
 export 'src/ui/app_bar/responsive_app_bar.dart';
 export 'src/ui/app_bar/widgets/app_bar_action.dart';