+++ lib/src/domain/entities/responsive_settings.dart
@@ -1,37 +1,14 @@
-import 'package:flutter/material.dart';
-import 'package:responsive_design/src/domain/enums/responsive_type.dart';
+import 'package:responsive_design/src/domain/entities/screen_change_points.dart';
 
 /// Manage the responsive screen settings.
 class ResponsiveSettings {
-  /// Define a [ResponsiveSettings] class.
-  const ResponsiveSettings({
-    this.desktopChangePoint = 1200.0,
-    this.tabletChangePoint = 600.0,
-    this.watchChangePoint = 300.0,
-  });
+  ResponsiveSettings._();
+  ScreenChangePoints changePoints = const ScreenChangePoints();
 
-  /// When the width is greater als this value
-  /// the display will be set as [ResponsiveType.desktop]
-  final double desktopChangePoint;
+  static ResponsiveSettings? _instance;
 
-  /// When the width is greater als this value
-  /// the display will be set as [ResponsiveType.tablet]
-  /// or when width greater als [watchChangePoint] and smaller als this value
-  /// the display will be [ResponsiveType.phone]
-  final double tabletChangePoint;
-
-  /// When the width is smaller als this value
-  /// the display will be set as [ResponsiveType.watch]
-  /// or when width greater als this value and smaller als [tabletChangePoint]
-  /// the display will be [ResponsiveType.phone]
-  final double watchChangePoint;
-
-  /// Returns the [ResponsiveType] by the screen size, calculated with [context].
-  ResponsiveType responsiveType(BuildContext context) {
-    final double width = MediaQuery.of(context).size.width;
-    if (width >= desktopChangePoint) return ResponsiveType.desktop;
-    if (width >= tabletChangePoint) return ResponsiveType.tablet;
-    if (width < watchChangePoint) return ResponsiveType.watch;
-    return ResponsiveType.phone;
+  // ignore: prefer_constructors_over_static_methods
+  static ResponsiveSettings get instance {
+    return _instance ??= ResponsiveSettings._();
   }
 }